Broke interface into function signatures, and moved types to seperate state file
